+++ backend/src/execution.rs
@@ -1,15 +1,20 @@
-mod executors;
 mod template_engine;
+mod util;
 mod wrappers;
 
+use crate::configuration::language_config::LanguageConfig;
 use crate::configuration::{get_config_for_lang, get_temp_dir};
 use crate::doc::TangleError;
 use crate::doc::TanglitDoc;
 use crate::errors::ExecutionError;
+use crate::execution::util::find_file_in_dir;
 use executors::execute_file;
 use log::debug;
-use std::process::Output;
+use std::path::Path;
+use std::process::{Command, Output, Stdio};
 pub use wrappers::{make_executable_code, write_file};
+
+const EXECUTE_SCRIPT_FILENAME: &str = "execute";
 
 /// Executes a code block by tangling it and adding necessary wrappers to make it executable.
 /// Prints both the resulting stdout and sterr from the execution and returns the stdout as a String.
@@ -50,12 +55,22 @@ pub fn execute(doc: &TanglitDoc, target_block: &str) -> Result<Output, Execution
 
     debug!("Wrote tangled code to file: {}", block_file_path.display());
 
-    let execution_script_path =
-        lang_config
-            .get_execution_script_path()
-            .ok_or(ExecutionError::InternalError(
-                "Execution script not found".to_string(),
-            ))?;
+    execute_block(&block_file_path, &lang_config)
+}
 
-    execute_file(&execution_script_path, &block_file_path)
+pub fn execute_block(
+    block_file_path: &Path,
+    lang_config: &LanguageConfig,
+) -> Result<Output, ExecutionError> {
+    let execution_script_path = find_file_in_dir(&lang_config.config_dir, EXECUTE_SCRIPT_FILENAME)
+        .ok_or(ExecutionError::InternalError(
+            "Execution script not found".to_string(),
+        ))?;
+
+    Command::new(execution_script_path)
+        .arg(block_file_path)
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .output()
+        .map_err(|e| ExecutionError::InternalError(e.to_string()))
 }
